// src/models/IoK8sApiNetworkingV1alpha1IPAddress/types.ts
import { IoK8sApiNetworkingV1alpha1IPAddressSpecModelOptions as IoK8sApiNetworkingV1alpha1IPAddressSpec } from '../IoK8sApiNetworkingV1alpha1IPAddressSpec';
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiNetworkingV1alpha1IPAddressModelOptions {
  apiVersion?: string;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  spec?: IoK8sApiNetworkingV1alpha1IPAddressSpec;
}