// src/models/IoK8sApiNetworkingV1alpha1IPAddressList/types.ts
import { IoK8sApiNetworkingV1alpha1IPAddressModelOptions as IoK8sApiNetworkingV1alpha1IPAddress } from '../IoK8sApiNetworkingV1alpha1IPAddress';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiNetworkingV1alpha1IPAddressListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiNetworkingV1alpha1IPAddress>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}